media_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
requests
PyJWT
boto3
firebase-admin
